JNG-4774 Check all type of transfer object is successfully created
